+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # docs
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ultify/datastructure@0.2.2
+
 ## 0.0.1
 
 ### Patch Changes